requirements.txt
genanki
google-genai
dot-env